refactor(frontend): Add LintService
