--- frontend/src/app/model/codegen/position.ts
@@ -1,4 +1,0 @@
-export interface Position {
-  line: number;
-  ch: number;
-}
