+++ frontend/src/app/scenario-editor.service.ts
@@ -1,13 +1,11 @@
-import {Injectable} from '@angular/core';
 import {HttpClient} from '@angular/common/http';
+import {Injectable} from '@angular/core';
 import {Observable} from 'rxjs';
 
-import {Marker} from './model/codegen/marker';
+import {environment} from '../environments/environment';
 import Request from './model/codegen/request';
 import Response from './model/codegen/response';
 import {PrivacyService} from './privacy.service';
-
-import {environment} from '../environments/environment';
 
 
 @Injectable({
@@ -22,32 +20,6 @@ export class ScenarioEditorService {
 
   submit(codeGenRequest: Request): Observable<Response> {
     return this.http.post<Response>(environment.apiURL + '/runcodegen', codeGenRequest);
-  }
-
-  lint(response: Response): Marker[] {
-    const result: Marker[] = [];
-
-    for (const line of response.output.split('\n')) {
-      const match = /^.*\.md:(\d+):(\d+)(?:-(\d+))?: (error|syntax|warning|note): (.*)$/.exec(line);
-      if (!match) {
-        continue;
-      }
-
-      const row = +match[1] - 1;
-      const col = +match[2];
-      const endCol = +(match[3] || col) + 1;
-      const severity = match[4] === 'syntax' ? 'error' : match[4];
-      const message = match[5];
-
-      result.push({
-        severity,
-        message,
-        from: {line: row, ch: col},
-        to: {line: row, ch: endCol},
-      });
-    }
-
-    return result;
   }
 
   foldInternalCalls(packageName: string, outputLines: string[]): string[] {
